move withLogger back into index
